requirements.txt
et-xmlfile==1.1.0
numpy==1.21.4
opencv-python==4.5.4.58
openpyxl==3.0.9
pdf2image==1.16.0
Pillow==8.4.0
PyPDF2==1.26.0
pytesseract==0.3.8
